% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\name{coefplot.newblblm}
\alias{coefplot.newblblm}
\title{Plot Boostrap Coefficient Estimates}
\usage{
\method{coefplot}{newblblm}(object, violin = FALSE, boots = FALSE, box = FALSE)
}
\arguments{
\item{object}{a newblblm class object}

\item{violin}{logical. Option to show a violin plot of the distribution of the bootstrap estimates.}

\item{boots}{logical. Option to show the estimation from each bootstrap.}

\item{box}{logical. Option to show a boxplot for the distribution of the bootstrap estimates.}
}
\description{
Provides a variety of plot to visualize the bootstrap estimates.
}
