% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\name{coefplot}
\alias{coefplot}
\title{Generic function coefplot}
\usage{
coefplot(object, violin = FALSE, boots = FALSE, box = FALSE)
}
\arguments{
\item{object}{the model}

\item{violin}{logical. Option to show a violin plot.}

\item{boots}{logical. Option to show a scatter plot.}

\item{box}{logical. Option to show a boxplot.}
}
\description{
Generic function coefplot
}
