% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\name{fitplot.newblblm}
\alias{fitplot.newblblm}
\title{Plot a newblblm Linear Model}
\usage{
\method{fitplot}{newblblm}(object, data, boots = FALSE)
}
\arguments{
\item{object}{a newblblm class object}

\item{data}{data.frame. The original data that the model is fitted on}

\item{boots}{logical. The option to show the linear regression line for each bootstrap}
}
\description{
Plot the data alongside the linear regression line computed from the newblblm procdure.
}
