% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\name{fitplot}
\alias{fitplot}
\title{Generic function fitplot}
\usage{
fitplot(object, data, boots = FALSE)
}
\arguments{
\item{object}{the model}

\item{data}{data.frame to be plotted}

\item{boots}{logical}
}
\description{
Generic function fitplot
}
