% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\name{future_estimate}
\alias{future_estimate}
\title{find the estimates}
\usage{
future_estimate(formula, data, m, B)
}
\arguments{
\item{formula}{a symbol description of the model to be fitted}

\item{data}{a data.frame of a vector list of file names}

\item{m}{integer. the number of subsamples}

\item{B}{integer. the number of bootstraps for each subsample}
}
\description{
find the estimates
}
