% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\name{lm1}
\alias{lm1}
\title{estimate the regression estimates based on given the number of repetitions}
\usage{
lm1(formula, data, freqs)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted}

\item{data}{the data the model will be fitted on}

\item{freqs}{the frequency of each observation in the dataset}
}
\description{
estimate the regression estimates based on given the number of repetitions
}
