% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\name{lm_each_boot}
\alias{lm_each_boot}
\title{compute the regression estimates for a blb dataset}
\usage{
lm_each_boot(formula, data, n)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted}

\item{data}{the data the model will be fitted on}

\item{n}{the size of the original data}
}
\description{
compute the regression estimates for a blb dataset
}
