% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\name{lm_each_subsample}
\alias{lm_each_subsample}
\title{compute the estimates}
\usage{
lm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted}

\item{data}{the data the model will be fitted on}

\item{n}{the size of the original data}

\item{B}{the number of bootstraps to be created}
}
\description{
compute the estimates
}
