% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\docType{package}
\name{newblblm-package}
\title{newblblm: Bag of Little Bootstraps Linear Model}
\description{
Utilizes the Bag of Little Bootstraps procedure to fit a linear regression model. 
  Allows for parallel computing.
}
