% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\name{newblblm}
\alias{newblblm}
\title{Linear Regression with Little Bag of Bootstraps}
\usage{
newblblm(formula, data, m = 10, B = 5000)
}
\arguments{
\item{formula}{a formula: a symbolic description of the model to be fitted.}

\item{data}{a data.frame or a vector list of file names.}

\item{m}{an integer for the number of subsets you want your data to be split into.}

\item{B}{an integer for the number of bootstraps you would like to use on each subset.}
}
\value{
returns an object of class "newblblm".
The applicable generic functions include boxplot, coef, confint, plot, predict, print, and sigma.
}
\description{
Linear Regression with Little Bag of Bootstraps
}
\details{
This model uses blb to find a lm fit.
}
\examples{
fit <- newblblm(mpg ~ wt, data = mtcars, m = 3, B = 1000)
}
