% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\name{plotboots}
\alias{plotboots}
\title{plot the regression line for each bootstrap}
\usage{
plotboots(object, data)
}
\arguments{
\item{object}{a class newblblm object}

\item{data}{the original data to be plotted}
}
\description{
plot the regression line for each bootstrap
}
