% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newblblm.R
\name{split_data}
\alias{split_data}
\title{split data into m parts of approximated equal sizes}
\usage{
split_data(data, m)
}
\arguments{
\item{data}{data.frame to split}

\item{m}{integer. the number of subsamples to make}
}
\description{
split data into m parts of approximated equal sizes
}
